+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `provider` label to tenant cluster metrics.
+
 ## [1.0.2] - 2020-10-12
 
 - Drop monitoring of non-managed kiam pods.